move prefab related classes into scene-graph
