+++ cocos/core/data/instantiate.ts
@@ -31,7 +31,7 @@ import { CCObject, isCCObject } from './object';
 import { js } from '../utils/js';
 import { getError, warn } from '../platform/debug';
 import { legacyCC } from '../global-exports';
-import { Prefab } from '../../asset/assets/prefab';
+import { Prefab } from '../../scene-graph/prefab';
 import { Node } from '../../scene-graph/node';
 import { updateChildrenForDeserialize } from '../utils/jsb-utils';
 import { isCCClassOrFastDefined } from './class';
