--- cocos/core/utils/prefab/index.ts
@@ -1,2 +1,0 @@
-export * from './prefab-info';
-export * from './utils';
