+++ cocos/scene-graph/prefab/index.ts
@@ -1,0 +1,5 @@
+// To keep compatibility:
+// - prefab-info.ts is exported in utils.ts
+// - utils.ts is exported in prefab.ts
+// - so here just export prefab.ts
+export * from './prefab';
