+++ cocos/scene-graph/prefab/prefab.ts
@@ -29,11 +29,11 @@ import { SUPPORT_JIT, ALIPAY, RUNTIME_BASED } from 'internal:constants';
 import { compile } from '../../core/data/instantiate-jit';
 import { js, obsolete } from '../../core/utils/js';
 import { Enum } from '../../core/value-types';
-import { Asset } from './asset';
-import { Node } from '../../scene-graph/node';
+import { Asset } from '../../asset/assets/asset';
+import { Node } from '../node';
 import { legacyCC } from '../../core/global-exports';
 import { warnID } from '../../core/platform/debug';
-import * as utils from '../../core/utils/prefab';
+import * as utils from './utils';
 
 /**
  * @en An enumeration used with the [[Prefab.optimizationPolicy]] to specify how to optimize the instantiate operation.
